+++ portal.go
@@ -41,27 +41,30 @@ import (
 	"sync"
 	"time"
 
-	"github.com/chai2010/webp"
 	"github.com/rs/zerolog"
 	"github.com/tidwall/gjson"
 	"go.mau.fi/util/dbutil"
 	"go.mau.fi/util/exerrors"
 	"go.mau.fi/util/exmime"
 	"go.mau.fi/util/ffmpeg"
+	"go.mau.fi/util/jsontime"
 	"go.mau.fi/util/random"
 	"go.mau.fi/util/variationselector"
+	cwebp "go.mau.fi/webp"
 	"go.mau.fi/whatsmeow"
 	waProto "go.mau.fi/whatsmeow/binary/proto"
 	"go.mau.fi/whatsmeow/types"
 	"go.mau.fi/whatsmeow/types/events"
 	"golang.org/x/exp/slices"
 	"golang.org/x/image/draw"
+	"golang.org/x/image/webp"
 	"google.golang.org/protobuf/proto"
 	log "maunium.net/go/maulogger/v2"
 	"maunium.net/go/mautrix"
 	"maunium.net/go/mautrix/appservice"
 	"maunium.net/go/mautrix/bridge"
 	"maunium.net/go/mautrix/bridge/bridgeconfig"
+	"maunium.net/go/mautrix/bridge/status"
 	"maunium.net/go/mautrix/crypto/attachment"
 	"maunium.net/go/mautrix/event"
 	"maunium.net/go/mautrix/format"
@@ -282,10 +285,48 @@ type Portal struct {
 
 	mediaErrorCache map[types.MessageID]*FailedMediaMeta
 
+	galleryCache          []*event.MessageEventContent
+	galleryCacheRootEvent id.EventID
+	galleryCacheStart     time.Time
+	galleryCacheReplyTo   *ReplyInfo
+	galleryCacheSender    types.JID
+
 	currentlySleepingToDelete sync.Map
 
 	relayUser    *User
 	parentPortal *Portal
+}
+
+const GalleryMaxTime = 10 * time.Minute
+
+func (portal *Portal) stopGallery() {
+	if portal.galleryCache != nil {
+		portal.galleryCache = nil
+		portal.galleryCacheSender = types.EmptyJID
+		portal.galleryCacheReplyTo = nil
+		portal.galleryCacheStart = time.Time{}
+		portal.galleryCacheRootEvent = ""
+	}
+}
+
+func (portal *Portal) startGallery(evt *events.Message, msg *ConvertedMessage) {
+	portal.galleryCache = []*event.MessageEventContent{msg.Content}
+	portal.galleryCacheSender = evt.Info.Sender.ToNonAD()
+	portal.galleryCacheReplyTo = msg.ReplyTo
+	portal.galleryCacheStart = time.Now()
+}
+
+func (portal *Portal) extendGallery(msg *ConvertedMessage) int {
+	portal.galleryCache = append(portal.galleryCache, msg.Content)
+	msg.Content = &event.MessageEventContent{
+		MsgType:             event.MsgBeeperGallery,
+		Body:                "Sent a gallery",
+		BeeperGalleryImages: portal.galleryCache,
+	}
+	msg.Content.SetEdit(portal.galleryCacheRootEvent)
+	// Don't set the gallery images in the edit fallback
+	msg.Content.BeeperGalleryImages = nil
+	return len(portal.galleryCache) - 1
 }
 
 var (
@@ -296,14 +337,14 @@ var (
 	_ bridge.TypingPortal              = (*Portal)(nil)
 )
 
-func (portal *Portal) handleMessageLoopItem(msg PortalMessage) {
+func (portal *Portal) handleWhatsAppMessageLoopItem(msg PortalMessage) {
 	if len(portal.MXID) == 0 {
 		if msg.fake == nil && msg.undecryptable == nil && (msg.evt == nil || !containsSupportedMessage(msg.evt.Message)) {
 			portal.log.Debugln("Not creating portal room for incoming message: message is not a chat message")
 			return
 		}
 		portal.log.Debugln("Creating Matrix room from incoming message")
-		err := portal.CreateMatrixRoom(msg.source, nil, false, true)
+		err := portal.CreateMatrixRoom(msg.source, nil, nil, false, true)
 		if err != nil {
 			portal.log.Errorln("Failed to create portal room:", err)
 			return
@@ -317,8 +358,10 @@ func (portal *Portal) handleMessageLoopItem(msg PortalMessage) {
 	case msg.receipt != nil:
 		portal.handleReceipt(msg.receipt, msg.source)
 	case msg.undecryptable != nil:
+		portal.stopGallery()
 		portal.handleUndecryptableMessage(msg.source, msg.undecryptable)
 	case msg.fake != nil:
+		portal.stopGallery()
 		msg.fake.ID = "FAKE::" + msg.fake.ID
 		portal.handleFakeMessage(*msg.fake)
 	default:
@@ -351,9 +394,36 @@ func (portal *Portal) handleMatrixMessageLoopItem(msg PortalMatrixMessage) {
 	}
 }
 
+func (portal *Portal) handleDeliveryReceipt(receipt *events.Receipt, source *User) {
+	if !portal.IsPrivateChat() {
+		return
+	}
+	for _, msgID := range receipt.MessageIDs {
+		msg := portal.bridge.DB.Message.GetByJID(portal.Key, msgID)
+		if msg == nil || msg.IsFakeMXID() {
+			continue
+		}
+		if msg.Sender == source.JID {
+			portal.bridge.SendRawMessageCheckpoint(&status.MessageCheckpoint{
+				EventID:    msg.MXID,
+				RoomID:     portal.MXID,
+				Step:       status.MsgStepRemote,
+				Timestamp:  jsontime.UM(receipt.Timestamp),
+				Status:     status.MsgStatusDelivered,
+				ReportedBy: status.MsgReportedByBridge,
+			})
+			portal.sendStatusEvent(msg.MXID, "", nil, &[]id.UserID{portal.MainIntent().UserID})
+		}
+	}
+}
+
 func (portal *Portal) handleReceipt(receipt *events.Receipt, source *User) {
 	if receipt.Sender.Server != types.DefaultUserServer {
 		// TODO handle lids
+		return
+	}
+	if receipt.Type == events.ReceiptTypeDelivered {
+		portal.handleDeliveryReceipt(receipt, source)
 		return
 	}
 	// The order of the message ID array depends on the sender's platform, so we just have to find
@@ -394,14 +464,31 @@ func (portal *Portal) handleReceipt(receipt *events.Receipt, source *User) {
 
 func (portal *Portal) handleMessageLoop() {
 	for {
-		select {
-		case msg := <-portal.messages:
-			portal.handleMessageLoopItem(msg)
-		case msg := <-portal.matrixMessages:
-			portal.handleMatrixMessageLoopItem(msg)
-		case retry := <-portal.mediaRetries:
-			portal.handleMediaRetry(retry.evt, retry.source)
+		portal.handleOneMessageLoopItem()
+	}
+}
+
+func (portal *Portal) handleOneMessageLoopItem() {
+	defer func() {
+		if err := recover(); err != nil {
+			logEvt := portal.zlog.WithLevel(zerolog.FatalLevel).
+				Str(zerolog.ErrorStackFieldName, string(debug.Stack()))
+			actualErr, ok := err.(error)
+			if ok {
+				logEvt = logEvt.Err(actualErr)
+			} else {
+				logEvt = logEvt.Any(zerolog.ErrorFieldName, err)
+			}
+			logEvt.Msg("Portal message handler panicked")
 		}
+	}()
+	select {
+	case msg := <-portal.messages:
+		portal.handleWhatsAppMessageLoopItem(msg)
+	case msg := <-portal.matrixMessages:
+		portal.handleMatrixMessageLoopItem(msg)
+	case retry := <-portal.mediaRetries:
+		portal.handleMediaRetry(retry.evt, retry.source)
 	}
 }
 
@@ -686,7 +773,7 @@ func (portal *Portal) handleUndecryptableMessage(source *User, evt *events.Undec
 	if evt.IsUnavailable {
 		metricType = "unavailable"
 	}
-	Segment.Track(source.MXID, "WhatsApp undecryptable message", map[string]interface{}{
+	Analytics.Track(source.MXID, "WhatsApp undecryptable message", map[string]interface{}{
 		"messageID":         evt.Info.ID,
 		"undecryptableType": metricType,
 	})
@@ -700,7 +787,7 @@ func (portal *Portal) handleUndecryptableMessage(source *User, evt *events.Undec
 		portal.log.Errorfln("Failed to send decryption error of %s to Matrix: %v", evt.Info.ID, err)
 		return
 	}
-	portal.finishHandling(nil, &evt.Info, resp.EventID, intent.UserID, database.MsgUnknown, database.MsgErrDecryptionFailed)
+	portal.finishHandling(nil, &evt.Info, resp.EventID, intent.UserID, database.MsgUnknown, 0, database.MsgErrDecryptionFailed)
 }
 
 func (portal *Portal) handleFakeMessage(msg fakeMessage) {
@@ -738,7 +825,7 @@ func (portal *Portal) handleFakeMessage(msg fakeMessage) {
 			MessageSource: types.MessageSource{
 				Sender: msg.Sender,
 			},
-		}, resp.EventID, intent.UserID, database.MsgFake, database.MsgNoError)
+		}, resp.EventID, intent.UserID, database.MsgFake, 0, database.MsgNoError)
 	}
 }
 
@@ -762,7 +849,7 @@ func (portal *Portal) handleMessage(source *User, evt *events.Message, historica
 			if evt.UnavailableRequestID != "" {
 				resolveType = "phone"
 			}
-			Segment.Track(source.MXID, "WhatsApp undecryptable message resolved", map[string]interface{}{
+			Analytics.Track(source.MXID, "WhatsApp undecryptable message resolved", map[string]interface{}{
 				"messageID":   evt.Info.ID,
 				"resolveType": resolveType,
 			})
@@ -795,6 +882,17 @@ func (portal *Portal) handleMessage(source *User, evt *events.Message, historica
 	}
 	converted := portal.convertMessage(intent, source, &evt.Info, evt.Message, false)
 	if converted != nil {
+		isGalleriable := portal.bridge.Config.Bridge.BeeperGalleries &&
+			(evt.Message.ImageMessage != nil || evt.Message.VideoMessage != nil) &&
+			(portal.galleryCache == nil ||
+				(evt.Info.Sender.ToNonAD() == portal.galleryCacheSender &&
+					converted.ReplyTo.Equals(portal.galleryCacheReplyTo) &&
+					time.Since(portal.galleryCacheStart) < GalleryMaxTime)) &&
+			// Captions aren't allowed in galleries (this needs to be checked before the caption is merged)
+			converted.Caption == nil &&
+			// Images can't be edited
+			editTargetMsg == nil
+
 		if !historical && portal.IsPrivateChat() && evt.Info.Sender.Device == 0 && converted.ExpiresIn > 0 && portal.ExpirationTime == 0 {
 			portal.zlog.Info().
 				Str("timer", converted.ExpiresIn.String()).
@@ -825,6 +923,20 @@ func (portal *Portal) handleMessage(source *User, evt *events.Message, historica
 			dbMsgType = database.MsgEdit
 			converted.Content.SetEdit(editTargetMsg.MXID)
 		}
+		galleryStarted := false
+		var galleryPart int
+		if isGalleriable {
+			if portal.galleryCache == nil {
+				portal.startGallery(evt, converted)
+				galleryStarted = true
+			} else {
+				galleryPart = portal.extendGallery(converted)
+				dbMsgType = database.MsgBeeperGallery
+			}
+		} else if editTargetMsg == nil {
+			// Stop collecting a gallery (except if it's an edit)
+			portal.stopGallery()
+		}
 		resp, err := portal.sendMessage(converted.Intent, converted.Type, converted.Content, converted.Extra, evt.Info.Timestamp.UnixMilli())
 		if err != nil {
 			portal.log.Errorfln("Failed to send %s to Matrix: %v", msgID, err)
@@ -834,6 +946,11 @@ func (portal *Portal) handleMessage(source *User, evt *events.Message, historica
 			}
 			eventID = resp.EventID
 			lastEventID = eventID
+			if galleryStarted {
+				portal.galleryCacheRootEvent = eventID
+			} else if galleryPart != 0 {
+				eventID = portal.galleryCacheRootEvent
+			}
 		}
 		// TODO figure out how to handle captions with undecryptable messages turning decryptable
 		if converted.Caption != nil && existingMsg == nil && editTargetMsg == nil {
@@ -866,7 +983,7 @@ func (portal *Portal) handleMessage(source *User, evt *events.Message, historica
 			}
 		}
 		if len(eventID) != 0 {
-			portal.finishHandling(existingMsg, &evt.Info, eventID, intent.UserID, dbMsgType, converted.Error)
+			portal.finishHandling(existingMsg, &evt.Info, eventID, intent.UserID, dbMsgType, galleryPart, converted.Error)
 		}
 	} else if msgType == "reaction" || msgType == "encrypted reaction" {
 		if evt.Message.GetEncReactionMessage() != nil {
@@ -911,12 +1028,13 @@ func (portal *Portal) isRecentlyHandled(id types.MessageID, error database.Messa
 	return false
 }
 
-func (portal *Portal) markHandled(txn dbutil.Transaction, msg *database.Message, info *types.MessageInfo, mxid id.EventID, senderMXID id.UserID, isSent, recent bool, msgType database.MessageType, errType database.MessageErrorType) *database.Message {
+func (portal *Portal) markHandled(txn dbutil.Transaction, msg *database.Message, info *types.MessageInfo, mxid id.EventID, senderMXID id.UserID, isSent, recent bool, msgType database.MessageType, galleryPart int, errType database.MessageErrorType) *database.Message {
 	if msg == nil {
 		msg = portal.bridge.DB.Message.New()
 		msg.Chat = portal.Key
 		msg.JID = info.ID
 		msg.MXID = mxid
+		msg.GalleryPart = galleryPart
 		msg.Timestamp = info.Timestamp
 		msg.Sender = info.Sender
 		msg.SenderMXID = senderMXID
@@ -959,6 +1077,9 @@ func (portal *Portal) getMessagePuppet(user *User, info *types.MessageInfo) (pup
 }
 
 func (portal *Portal) getMessageIntent(user *User, info *types.MessageInfo, msgType string) *appservice.IntentAPI {
+	if portal.IsNewsletter() && info.Sender == info.Chat {
+		return portal.MainIntent()
+	}
 	puppet := portal.getMessagePuppet(user, info)
 	if puppet == nil {
 		return nil
@@ -971,8 +1092,8 @@ func (portal *Portal) getMessageIntent(user *User, info *types.MessageInfo, msgT
 	return intent
 }
 
-func (portal *Portal) finishHandling(existing *database.Message, message *types.MessageInfo, mxid id.EventID, senderMXID id.UserID, msgType database.MessageType, errType database.MessageErrorType) {
-	portal.markHandled(nil, existing, message, mxid, senderMXID, true, true, msgType, errType)
+func (portal *Portal) finishHandling(existing *database.Message, message *types.MessageInfo, mxid id.EventID, senderMXID id.UserID, msgType database.MessageType, galleryPart int, errType database.MessageErrorType) {
+	portal.markHandled(nil, existing, message, mxid, senderMXID, true, true, msgType, galleryPart, errType)
 	portal.sendDeliveryReceipt(mxid)
 	var suffix string
 	if errType == database.MsgErrDecryptionFailed {
@@ -1043,6 +1164,9 @@ func (portal *Portal) syncParticipant(source *User, participant types.GroupParti
 }
 
 func (portal *Portal) SyncParticipants(source *User, metadata *types.GroupInfo) ([]id.UserID, *event.PowerLevelsEventContent) {
+	if portal.IsNewsletter() {
+		return nil, nil
+	}
 	changed := false
 	var levels *event.PowerLevelsEventContent
 	var err error
@@ -1121,6 +1245,18 @@ func reuploadAvatar(intent *appservice.IntentAPI, url string) (id.ContentURI, er
 	return resp.ContentURI, nil
 }
 
+func (user *User) reuploadAvatarDirectPath(intent *appservice.IntentAPI, directPath string) (id.ContentURI, error) {
+	data, err := user.Client.DownloadMediaWithPath(directPath, nil, nil, nil, 0, "", "")
+	if err != nil {
+		return id.ContentURI{}, fmt.Errorf("failed to download avatar: %w", err)
+	}
+	resp, err := intent.UploadBytes(data, http.DetectContentType(data))
+	if err != nil {
+		return id.ContentURI{}, fmt.Errorf("failed to upload avatar to Matrix: %w", err)
+	}
+	return resp.ContentURI, nil
+}
+
 func (user *User) updateAvatar(jid types.JID, isCommunity bool, avatarID *string, avatarURL *id.ContentURI, avatarSet *bool, log log.Logger, intent *appservice.IntentAPI) bool {
 	currentID := ""
 	if *avatarSet && *avatarID != "remove" && *avatarID != "unauthorized" {
@@ -1155,14 +1291,23 @@ func (user *User) updateAvatar(jid types.JID, isCommunity bool, avatarID *string
 	}
 	if avatar.ID == *avatarID && *avatarSet {
 		return false
-	} else if len(avatar.URL) == 0 {
+	} else if len(avatar.URL) == 0 && len(avatar.DirectPath) == 0 {
 		log.Warnln("Didn't get URL in response to avatar query")
 		return false
 	} else if avatar.ID != *avatarID || avatarURL.IsEmpty() {
-		url, err := reuploadAvatar(intent, avatar.URL)
-		if err != nil {
-			log.Warnln("Failed to reupload avatar:", err)
-			return false
+		var url id.ContentURI
+		if len(avatar.URL) > 0 {
+			url, err = reuploadAvatar(intent, avatar.URL)
+			if err != nil {
+				log.Warnln("Failed to reupload avatar:", err)
+				return false
+			}
+		} else {
+			url, err = user.reuploadAvatarDirectPath(intent, avatar.DirectPath)
+			if err != nil {
+				log.Warnln("Failed to reupload avatar:", err)
+				return false
+			}
 		}
 		*avatarURL = url
 	}
@@ -1172,10 +1317,61 @@ func (user *User) updateAvatar(jid types.JID, isCommunity bool, avatarID *string
 	return true
 }
 
+func (portal *Portal) UpdateNewsletterAvatar(user *User, meta *types.NewsletterMetadata) bool {
+	portal.avatarLock.Lock()
+	defer portal.avatarLock.Unlock()
+	var picID string
+	picture := meta.ThreadMeta.Picture
+	if picture == nil {
+		picID = meta.ThreadMeta.Preview.ID
+	} else {
+		picID = picture.ID
+	}
+	if picID == "" {
+		picID = "remove"
+	}
+	if portal.Avatar != picID || !portal.AvatarSet {
+		if picID == "remove" {
+			portal.AvatarURL = id.ContentURI{}
+		} else if portal.Avatar != picID || portal.AvatarURL.IsEmpty() {
+			var err error
+			if picture == nil {
+				meta, err = user.Client.GetNewsletterInfo(portal.Key.JID)
+				if err != nil {
+					portal.log.Warnln("Failed to fetch full res avatar info for newsletter:", err)
+					return false
+				}
+				picture = meta.ThreadMeta.Picture
+				if picture == nil {
+					portal.log.Warnln("Didn't get full res avatar info for newsletter")
+					return false
+				}
+				picID = picture.ID
+			}
+			portal.AvatarURL, err = user.reuploadAvatarDirectPath(portal.MainIntent(), picture.DirectPath)
+			if err != nil {
+				portal.log.Warnln("Failed to reupload newsletter avatar:", err)
+				return false
+			}
+		}
+		portal.Avatar = picID
+		portal.AvatarSet = false
+		return portal.setRoomAvatar(true, types.EmptyJID, true)
+	}
+	return false
+}
+
 func (portal *Portal) UpdateAvatar(user *User, setBy types.JID, updateInfo bool) bool {
+	if portal.IsNewsletter() {
+		return false
+	}
 	portal.avatarLock.Lock()
 	defer portal.avatarLock.Unlock()
 	changed := user.updateAvatar(portal.Key.JID, portal.IsParent, &portal.Avatar, &portal.AvatarURL, &portal.AvatarSet, portal.log, portal.MainIntent())
+	return portal.setRoomAvatar(changed, setBy, updateInfo)
+}
+
+func (portal *Portal) setRoomAvatar(changed bool, setBy types.JID, updateInfo bool) bool {
 	if !changed || portal.Avatar == "unauthorized" {
 		if changed || updateInfo {
 			portal.Update(nil)
@@ -1273,6 +1469,21 @@ func (portal *Portal) UpdateTopic(topic string, setBy types.JID, updateInfo bool
 	return false
 }
 
+func newsletterToGroupInfo(meta *types.NewsletterMetadata) *types.GroupInfo {
+	var out types.GroupInfo
+	out.JID = meta.ID
+	out.Name = meta.ThreadMeta.Name.Text
+	out.NameSetAt = meta.ThreadMeta.Name.UpdateTime.Time
+	out.Topic = meta.ThreadMeta.Description.Text
+	out.TopicSetAt = meta.ThreadMeta.Description.UpdateTime.Time
+	out.TopicID = meta.ThreadMeta.Description.ID
+	out.GroupCreated = meta.ThreadMeta.CreationTime.Time
+	out.IsAnnounce = true
+	out.IsLocked = true
+	out.IsIncognito = true
+	return &out
+}
+
 func (portal *Portal) UpdateParentGroup(source *User, parent types.JID, updateInfo bool) bool {
 	portal.parentGroupUpdateLock.Lock()
 	defer portal.parentGroupUpdateLock.Unlock()
@@ -1294,7 +1505,7 @@ func (portal *Portal) UpdateParentGroup(source *User, parent types.JID, updateIn
 	return false
 }
 
-func (portal *Portal) UpdateMetadata(user *User, groupInfo *types.GroupInfo) bool {
+func (portal *Portal) UpdateMetadata(user *User, groupInfo *types.GroupInfo, newsletterMetadata *types.NewsletterMetadata) bool {
 	if portal.IsPrivateChat() {
 		return false
 	} else if portal.IsStatusBroadcastList() {
@@ -1316,6 +1527,17 @@ func (portal *Portal) UpdateMetadata(user *User, groupInfo *types.GroupInfo) boo
 		//}
 		//update = portal.UpdateTopic(BroadcastTopic, "", nil, false) || update
 		return update
+	}
+	if groupInfo == nil && portal.IsNewsletter() {
+		if newsletterMetadata == nil {
+			var err error
+			newsletterMetadata, err = user.Client.GetNewsletterInfo(portal.Key.JID)
+			if err != nil {
+				portal.zlog.Err(err).Msg("Failed to get newsletter info")
+				return false
+			}
+		}
+		groupInfo = newsletterToGroupInfo(newsletterMetadata)
 	}
 	if groupInfo == nil {
 		var err error
@@ -1345,6 +1567,9 @@ func (portal *Portal) UpdateMetadata(user *User, groupInfo *types.GroupInfo) boo
 
 	portal.RestrictMessageSending(groupInfo.IsAnnounce)
 	portal.RestrictMetadataChanges(groupInfo.IsLocked)
+	if newsletterMetadata != nil && newsletterMetadata.ViewerMeta != nil {
+		portal.PromoteNewsletterUser(user, newsletterMetadata.ViewerMeta.Role)
+	}
 
 	return update
 }
@@ -1353,7 +1578,7 @@ func (portal *Portal) ensureUserInvited(user *User) bool {
 	return user.ensureInvited(portal.MainIntent(), portal.MXID, portal.IsPrivateChat())
 }
 
-func (portal *Portal) UpdateMatrixRoom(user *User, groupInfo *types.GroupInfo) bool {
+func (portal *Portal) UpdateMatrixRoom(user *User, groupInfo *types.GroupInfo, newsletterMetadata *types.NewsletterMetadata) bool {
 	if len(portal.MXID) == 0 {
 		return false
 	}
@@ -1362,10 +1587,16 @@ func (portal *Portal) UpdateMatrixRoom(user *User, groupInfo *types.GroupInfo) b
 	portal.ensureUserInvited(user)
 	go portal.addToPersonalSpace(user)
 
+	if groupInfo == nil && newsletterMetadata != nil {
+		groupInfo = newsletterToGroupInfo(newsletterMetadata)
+	}
+
 	update := false
-	update = portal.UpdateMetadata(user, groupInfo) || update
-	if !portal.IsPrivateChat() && !portal.IsBroadcastList() {
+	update = portal.UpdateMetadata(user, groupInfo, newsletterMetadata) || update
+	if !portal.IsPrivateChat() && !portal.IsBroadcastList() && !portal.IsNewsletter() {
 		update = portal.UpdateAvatar(user, types.EmptyJID, false) || update
+	} else if newsletterMetadata != nil {
+		update = portal.UpdateNewsletterAvatar(user, newsletterMetadata) || update
 	}
 	if update || portal.LastSync.Add(24*time.Hour).Before(time.Now()) {
 		portal.LastSync = time.Now()
@@ -1463,6 +1694,35 @@ func (portal *Portal) RestrictMessageSending(restrict bool) id.EventID {
 	}
 
 	levels.EventsDefault = newLevel
+	resp, err := portal.MainIntent().SetPowerLevels(portal.MXID, levels)
+	if err != nil {
+		portal.log.Errorln("Failed to change power levels:", err)
+		return ""
+	} else {
+		return resp.EventID
+	}
+}
+
+func (portal *Portal) PromoteNewsletterUser(user *User, role types.NewsletterRole) id.EventID {
+	levels, err := portal.MainIntent().PowerLevels(portal.MXID)
+	if err != nil {
+		levels = portal.GetBasePowerLevels()
+	}
+
+	newLevel := 0
+	switch role {
+	case types.NewsletterRoleAdmin:
+		newLevel = 50
+	case types.NewsletterRoleOwner:
+		newLevel = 95
+	}
+
+	changed := portal.applyPowerLevelFixes(levels)
+	changed = levels.EnsureUserLevel(user.MXID, newLevel) || changed
+	if !changed {
+		return ""
+	}
+
 	resp, err := portal.MainIntent().SetPowerLevels(portal.MXID, levels)
 	if err != nil {
 		portal.log.Errorln("Failed to change power levels:", err)
@@ -1573,7 +1833,7 @@ func (portal *Portal) GetEncryptionEventContent() (evt *event.EncryptionEventCon
 	return
 }
 
-func (portal *Portal) CreateMatrixRoom(user *User, groupInfo *types.GroupInfo, isFullInfo, backfill bool) error {
+func (portal *Portal) CreateMatrixRoom(user *User, groupInfo *types.GroupInfo, newsletterMetadata *types.NewsletterMetadata, isFullInfo, backfill bool) error {
 	portal.roomCreateLock.Lock()
 	defer portal.roomCreateLock.Unlock()
 	if len(portal.MXID) > 0 {
@@ -1620,7 +1880,18 @@ func (portal *Portal) CreateMatrixRoom(user *User, groupInfo *types.GroupInfo, i
 		portal.log.Debugln("Broadcast list is not yet supported, not creating room after all")
 		return fmt.Errorf("broadcast list bridging is currently not supported")
 	} else {
-		if groupInfo == nil || !isFullInfo {
+		if portal.IsNewsletter() {
+			if newsletterMetadata == nil {
+				var err error
+				newsletterMetadata, err = user.Client.GetNewsletterInfo(portal.Key.JID)
+				if err != nil {
+					return err
+				}
+			}
+			if groupInfo == nil {
+				groupInfo = newsletterToGroupInfo(newsletterMetadata)
+			}
+		} else if groupInfo == nil || !isFullInfo {
 			foundInfo, err := user.Client.GetGroupInfo(portal.Key.JID)
 
 			// Ensure that the user is actually a participant in the conversation
@@ -1646,7 +1917,11 @@ func (portal *Portal) CreateMatrixRoom(user *User, groupInfo *types.GroupInfo, i
 				portal.ExpirationTime = groupInfo.DisappearingTimer
 			}
 		}
-		portal.UpdateAvatar(user, types.EmptyJID, false)
+		if portal.IsNewsletter() {
+			portal.UpdateNewsletterAvatar(user, newsletterMetadata)
+		} else {
+			portal.UpdateAvatar(user, types.EmptyJID, false)
+		}
 	}
 
 	powerLevels := portal.GetBasePowerLevels()
@@ -1659,6 +1934,14 @@ func (portal *Portal) CreateMatrixRoom(user *User, groupInfo *types.GroupInfo, i
 			powerLevels.EnsureEventLevel(event.StateRoomName, 50)
 			powerLevels.EnsureEventLevel(event.StateRoomAvatar, 50)
 			powerLevels.EnsureEventLevel(event.StateTopic, 50)
+		}
+	}
+	if newsletterMetadata != nil && newsletterMetadata.ViewerMeta != nil {
+		switch newsletterMetadata.ViewerMeta.Role {
+		case types.NewsletterRoleAdmin:
+			powerLevels.EnsureUserLevel(user.MXID, 50)
+		case types.NewsletterRoleOwner:
+			powerLevels.EnsureUserLevel(user.MXID, 95)
 		}
 	}
 
@@ -1725,7 +2008,7 @@ func (portal *Portal) CreateMatrixRoom(user *User, groupInfo *types.GroupInfo, i
 	autoJoinInvites := portal.bridge.SpecVersions.Supports(mautrix.BeeperFeatureAutojoinInvites)
 	if autoJoinInvites {
 		portal.log.Debugfln("Hungryserv mode: adding all group members in create request")
-		if groupInfo != nil {
+		if groupInfo != nil && !portal.IsNewsletter() {
 			// TODO non-hungryserv could also include all members in invites, and then send joins manually?
 			participants, powerLevels := portal.SyncParticipants(user, groupInfo)
 			invite = append(invite, participants...)
@@ -1793,7 +2076,7 @@ func (portal *Portal) CreateMatrixRoom(user *User, groupInfo *types.GroupInfo, i
 	go portal.updateCommunitySpace(user, true, true)
 	go portal.addToPersonalSpace(user)
 
-	if groupInfo != nil && !autoJoinInvites {
+	if !portal.IsNewsletter() && groupInfo != nil && !autoJoinInvites {
 		portal.SyncParticipants(user, groupInfo)
 	}
 	//if broadcastMetadata != nil {
@@ -1871,7 +2154,7 @@ func (portal *Portal) updateCommunitySpace(user *User, add, updateInfo bool) boo
 			return false
 		}
 		portal.log.Debugfln("Creating portal for parent group %v", space.Key.JID)
-		err := space.CreateMatrixRoom(user, nil, false, false)
+		err := space.CreateMatrixRoom(user, nil, nil, false, false)
 		if err != nil {
 			portal.log.Debugfln("Failed to create portal for parent group: %v", err)
 			return false
@@ -1919,6 +2202,10 @@ func (portal *Portal) IsGroupChat() bool {
 
 func (portal *Portal) IsBroadcastList() bool {
 	return portal.Key.JID.Server == types.BroadcastServer
+}
+
+func (portal *Portal) IsNewsletter() bool {
+	return portal.Key.JID.Server == types.NewsletterServer
 }
 
 func (portal *Portal) IsStatusBroadcastList() bool {
@@ -2054,7 +2341,7 @@ func (portal *Portal) HandleMessageReaction(intent *appservice.IntentAPI, user *
 		if err != nil {
 			portal.log.Errorfln("Failed to redact reaction %s/%s from %s to %s: %v", existing.MXID, existing.JID, info.Sender, targetJID, err)
 		}
-		portal.finishHandling(existingMsg, info, resp.EventID, intent.UserID, database.MsgReaction, database.MsgNoError)
+		portal.finishHandling(existingMsg, info, resp.EventID, intent.UserID, database.MsgReaction, 0, database.MsgNoError)
 		existing.Delete()
 	} else {
 		target := portal.bridge.DB.Message.GetByJID(portal.Key, targetJID)
@@ -2075,7 +2362,7 @@ func (portal *Portal) HandleMessageReaction(intent *appservice.IntentAPI, user *
 			return
 		}
 
-		portal.finishHandling(existingMsg, info, resp.EventID, intent.UserID, database.MsgReaction, database.MsgNoError)
+		portal.finishHandling(existingMsg, info, resp.EventID, intent.UserID, database.MsgReaction, 0, database.MsgNoError)
 		portal.upsertReaction(nil, intent, target.JID, info.Sender, resp.EventID, info.ID)
 	}
 }
@@ -2161,6 +2448,15 @@ type ReplyInfo struct {
 	MessageID types.MessageID
 	Chat      types.JID
 	Sender    types.JID
+}
+
+func (r *ReplyInfo) Equals(other *ReplyInfo) bool {
+	if r == nil {
+		return other == nil
+	} else if other == nil {
+		return false
+	}
+	return r.MessageID == other.MessageID && r.Chat == other.Chat && r.Sender == other.Sender
 }
 
 func (r ReplyInfo) MarshalZerologObject(e *zerolog.Event) {
@@ -3495,7 +3791,7 @@ func (portal *Portal) convertToWebP(img []byte) ([]byte, error) {
 	}
 
 	var webpBuffer bytes.Buffer
-	if err = webp.Encode(&webpBuffer, decodedImg, nil); err != nil {
+	if err = cwebp.Encode(&webpBuffer, decodedImg, nil); err != nil {
 		return img, fmt.Errorf("failed to encode webp image: %w", err)
 	}
 
@@ -3538,6 +3834,9 @@ func (portal *Portal) preprocessMatrixMedia(ctx context.Context, sender *User, r
 		}
 	}
 	mimeType := content.GetInfo().MimeType
+	if mimeType == "" {
+		content.Info.MimeType = "application/octet-stream"
+	}
 	var convertErr error
 	// Allowed mime types from https://developers.facebook.com/docs/whatsapp/on-premises/reference/media
 	switch {
@@ -3841,7 +4140,6 @@ func (portal *Portal) convertMatrixPollStart(_ context.Context, sender *User, ev
 	if maxAnswers >= len(content.PollStart.Answers) || maxAnswers < 0 {
 		maxAnswers = 0
 	}
-	fmt.Printf("%+v\n", content.PollStart)
 	ctxInfo := portal.generateContextInfo(content.RelatesTo)
 	var question string
 	question, ctxInfo.MentionedJid = portal.msc1767ToWhatsApp(content.PollStart.Question, true)
@@ -3882,7 +4180,7 @@ func (portal *Portal) generateContextInfo(relatesTo *event.RelatesTo) *waProto.C
 	replyToID := relatesTo.GetReplyTo()
 	if len(replyToID) > 0 {
 		replyToMsg := portal.bridge.DB.Message.GetByMXID(replyToID)
-		if replyToMsg != nil && !replyToMsg.IsFakeJID() && (replyToMsg.Type == database.MsgNormal || replyToMsg.Type == database.MsgMatrixPoll) {
+		if replyToMsg != nil && !replyToMsg.IsFakeJID() && (replyToMsg.Type == database.MsgNormal || replyToMsg.Type == database.MsgMatrixPoll || replyToMsg.Type == database.MsgBeeperGallery) {
 			ctxInfo.StanzaId = &replyToMsg.JID
 			ctxInfo.Participant = proto.String(replyToMsg.Sender.ToNonAD().String())
 			// Using blank content here seems to work fine on all official WhatsApp apps.
@@ -3902,6 +4200,8 @@ func (portal *Portal) generateContextInfo(relatesTo *event.RelatesTo) *waProto.C
 type extraConvertMeta struct {
 	PollOptions map[[32]byte]string
 	EditRootMsg *database.Message
+
+	GalleryExtraParts []*waProto.Message
 }
 
 func getEditError(rootMsg *database.Message, editer *User) error {
@@ -3937,6 +4237,9 @@ func (portal *Portal) convertMatrixMessage(ctx context.Context, sender *User, ev
 			return nil, sender, extraMeta, errUserNotLoggedIn
 		}
 		sender = portal.GetRelayUser()
+		if !sender.IsLoggedIn() {
+			return nil, sender, extraMeta, errRelaybotNotLoggedIn
+		}
 		isRelay = true
 	}
 	var editRootMsg *database.Message
@@ -4012,6 +4315,37 @@ func (portal *Portal) convertMatrixMessage(ctx context.Context, sender *User, ev
 			FileEncSha256: media.FileEncSHA256,
 			FileSha256:    media.FileSHA256,
 			FileLength:    proto.Uint64(uint64(media.FileLength)),
+		}
+	case event.MsgBeeperGallery:
+		if isRelay {
+			return nil, sender, extraMeta, errGalleryRelay
+		} else if content.BeeperGalleryCaption != "" {
+			return nil, sender, extraMeta, errGalleryCaption
+		}
+		for i, part := range content.BeeperGalleryImages {
+			// TODO support videos
+			media, err := portal.preprocessMatrixMedia(ctx, sender, false, part, evt.ID, whatsmeow.MediaImage)
+			if media == nil {
+				return nil, sender, extraMeta, fmt.Errorf("failed to handle image #%d: %w", i+1, err)
+			}
+			imageMsg := &waProto.ImageMessage{
+				ContextInfo:   ctxInfo,
+				JpegThumbnail: media.Thumbnail,
+				Url:           &media.URL,
+				DirectPath:    &media.DirectPath,
+				MediaKey:      media.MediaKey,
+				Mimetype:      &part.GetInfo().MimeType,
+				FileEncSha256: media.FileEncSHA256,
+				FileSha256:    media.FileSHA256,
+				FileLength:    proto.Uint64(uint64(media.FileLength)),
+			}
+			if i == 0 {
+				msg.ImageMessage = imageMsg
+			} else {
+				extraMeta.GalleryExtraParts = append(extraMeta.GalleryExtraParts, &waProto.Message{
+					ImageMessage: imageMsg,
+				})
+			}
 		}
 	case event.MessageType(event.EventSticker.Type):
 		media, err := portal.preprocessMatrixMedia(ctx, sender, relaybotFormatted, content, evt.ID, whatsmeow.MediaImage)
@@ -4238,7 +4572,7 @@ func (portal *Portal) HandleMatrixMessage(sender *User, evt *event.Event, timing
 	}
 	info := portal.generateMessageInfo(sender)
 	if dbMsg == nil {
-		dbMsg = portal.markHandled(nil, nil, info, evt.ID, evt.Sender, false, true, dbMsgType, database.MsgNoError)
+		dbMsg = portal.markHandled(nil, nil, info, evt.ID, evt.Sender, false, true, dbMsgType, 0, database.MsgNoError)
 	} else {
 		info.ID = dbMsg.JID
 	}
@@ -4250,10 +4584,26 @@ func (portal *Portal) HandleMatrixMessage(sender *User, evt *event.Event, timing
 	resp, err := sender.Client.SendMessage(ctx, portal.Key.JID, msg, whatsmeow.SendRequestExtra{ID: info.ID})
 	timings.totalSend = time.Since(start)
 	timings.whatsmeow = resp.DebugTimings
-	go ms.sendMessageMetrics(evt, err, "Error sending", true)
-	if err == nil {
-		dbMsg.MarkSent(resp.Timestamp)
+	if err != nil {
+		go ms.sendMessageMetrics(evt, err, "Error sending", true)
+		return
 	}
+	dbMsg.MarkSent(resp.Timestamp)
+	if len(extraMeta.GalleryExtraParts) > 0 {
+		for i, part := range extraMeta.GalleryExtraParts {
+			partInfo := portal.generateMessageInfo(sender)
+			partDBMsg := portal.markHandled(nil, nil, partInfo, evt.ID, evt.Sender, false, true, database.MsgBeeperGallery, i+1, database.MsgNoError)
+			portal.log.Debugln("Sending gallery part", i+2, "of event", evt.ID, "to WhatsApp", partInfo.ID)
+			resp, err = sender.Client.SendMessage(ctx, portal.Key.JID, part, whatsmeow.SendRequestExtra{ID: partInfo.ID})
+			if err != nil {
+				go ms.sendMessageMetrics(evt, err, "Error sending", true)
+				return
+			}
+			portal.log.Debugfln("Sent gallery part", i+2, "of event", evt.ID)
+			partDBMsg.MarkSent(resp.Timestamp)
+		}
+	}
+	go ms.sendMessageMetrics(evt, nil, "", true)
 }
 
 func (portal *Portal) HandleMatrixReaction(sender *User, evt *event.Event) {
@@ -4293,7 +4643,7 @@ func (portal *Portal) handleMatrixReaction(sender *User, evt *event.Event) error
 		return fmt.Errorf("unknown target event %s", content.RelatesTo.EventID)
 	}
 	info := portal.generateMessageInfo(sender)
-	dbMsg := portal.markHandled(nil, nil, info, evt.ID, evt.Sender, false, true, database.MsgReaction, database.MsgNoError)
+	dbMsg := portal.markHandled(nil, nil, info, evt.ID, evt.Sender, false, true, database.MsgReaction, 0, database.MsgNoError)
 	portal.upsertReaction(nil, nil, target.JID, sender.JID, evt.ID, info.ID)
 	portal.log.Debugln("Sending reaction", evt.ID, "to WhatsApp", info.ID)
 	resp, err := portal.sendReactionToWhatsApp(sender, info.ID, target, content.RelatesTo.Key, evt.Timestamp)
@@ -4309,7 +4659,9 @@ func (portal *Portal) sendReactionToWhatsApp(sender *User, id types.MessageID, t
 		messageKeyParticipant = proto.String(target.Sender.ToNonAD().String())
 	}
 	key = variationselector.Remove(key)
-	return sender.Client.SendMessage(context.TODO(), portal.Key.JID, &waProto.Message{
+	ctx, cancel := context.WithTimeout(context.TODO(), 60*time.Second)
+	defer cancel()
+	return sender.Client.SendMessage(ctx, portal.Key.JID, &waProto.Message{
 		ReactionMessage: &waProto.ReactionMessage{
 			Key: &waProto.MessageKey{
 				RemoteJid:   proto.String(portal.Key.JID.String()),
@@ -4395,7 +4747,9 @@ func (portal *Portal) HandleMatrixRedaction(sender *User, evt *event.Event) {
 			key.Participant = proto.String(msg.Sender.ToNonAD().String())
 		}
 		portal.log.Debugfln("Sending redaction %s of %s/%s to WhatsApp", evt.ID, msg.MXID, msg.JID)
-		_, err := sender.Client.SendMessage(context.TODO(), portal.Key.JID, &waProto.Message{
+		ctx, cancel := context.WithTimeout(context.TODO(), 60*time.Second)
+		defer cancel()
+		_, err := sender.Client.SendMessage(ctx, portal.Key.JID, &waProto.Message{
 			ProtocolMessage: &waProto.ProtocolMessage{
 				Type: waProto.ProtocolMessage_REVOKE.Enum(),
 				Key:  key,